requirements.txt
fastapi
uvicorn
neomodel
python-dotenv
gunicorn